Bugfix: Creating "data.aos" file when not founded.
Instead of relying on the error message - which is error prone - check
for the type of the error (FileTypeError), and then continue with the
procedure of creating the "data.aos" file for the first time.

I runned into this bug because the error message was different on my
OS due to my language, showing that this cold happen on the future, if
the formated error message changes on different python versions - which
is a bit unlikely to happen, but who knows?.

The best ever approach to this problem will be, in the future, create a
custom Exception, like AOSInitializationException(Exception), and maybe
a MissingDataAOSError(AOSInitializationException) exception, creating a
exception hierarchy that will make Exception Handling much more easy by
selecting just the desired error to handle.